+++ src/Za11.java
@@ -6,12 +6,9 @@ public class Za11 {
         String str = in.next();
         System.out.println("Ввести второе  число для int: ");
         int i = in.nextInt();
-        int max = Math.max(Integer.parseInt(str),  i);
-        double min = Math.min(Integer.parseInt(str),  i);;
+        double max = Math.max(Double.parseDouble(str),  i);
+        double min = Math.min(Double.parseDouble(str),  i);;
         System.out.println("Из введенного больше - "+ max);
         System.out.println("Из введенного меньше - "+ min);
-
-
-
     }
 }
